pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mariadb-bootstrap"
version = "0.1.0"
description = "Provision MariaDB databases/users from a TOML or YAML manifest"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "mysql-connector-python>=8.3",
    "PyYAML>=6.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
bootstrap-mariadb = "dborbit.cli:cli"

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]
